+++ src/app/shared/models/interfaces.ts
@@ -1,30 +1,7 @@
-import { AlertTypes } from './enums';
-
-export interface ILoginData {
-  email: string;
-  password: string;
-  returnSecureToken: boolean;
-}
-
-export interface IUser {
-  displayName: string;
-  email: string;
-  expiresIn: string;
-  idToken: string;
-  kind: string;
-  localId: string;
-  registered: boolean;
-}
-
 export interface IPost {
   id?: string;
   title: string;
   author: string;
   content?: string;
   date?: Date;
-}
-
-export interface IAlert {
-  type: AlertTypes;
-  text: string;
 }
