+++ src/app/shared/views/admin/admin-layout/admin-layout.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthService } from './../../../services/auth.service';
+import { AuthService } from '../shared/services/auth.service';
 @Component({
   selector: 'app-admin-layout',
   templateUrl: './admin-layout.component.html',
