+++ src/app/shared/views/admin/admin.module.ts
@@ -1,4 +1,4 @@
-import { SearchPostsPipe } from './../../pipes/search-posts.pipe';
+import { SearchPostsPipe } from './shared/pipes/search-posts.pipe';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
@@ -6,16 +6,14 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { SharedModule } from './../../shared.module';
 
-import { AuthService } from './../../services/auth.service';
-
-import { AuthGuard } from './../../guards/auth.guard';
+import { AuthGuard } from './shared/guards/auth.guard';
 
 import { AdminLayoutComponent } from './admin-layout/admin-layout.component';
 import { LoginPageComponent } from './login-page/login-page.component';
 import { EditPageComponent } from './edit-page/edit-page.component';
 import { DashboardPageComponent } from './dashboard-page/dashboard-page.component';
 import { CreatePageComponent } from './create-page/create-page.component';
-import { AlertComponent } from 'src/app/components/alert/alert.component';
+import { AlertComponent } from 'src/app/shared/views/admin/shared/components/alert/alert.component';
 
 const adminChildrens = [
   { path: '', redirectTo: '/admin/login', pathMatch: 'full' },
