+++ src/app/shared/views/admin/dashboard-page/dashboard-page.component.ts
@@ -3,11 +3,11 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 
 import { PostsService } from './../../../services/posts.service';
-import { AlertService } from 'src/app/shared/services/alert.service';
+import { AlertService } from '../shared/services/alert.service';
 
 import { IPost } from './../../../models/interfaces';
 
-import { AlertTypes } from 'src/app/shared/models/enums';
+import { AlertTypes } from './../shared/models/enums';
 @Component({
   selector: 'app-dashboard-page',
   templateUrl: './dashboard-page.component.html',
