+++ src/app/shared/views/admin/shared/components/alert/alert.component.scss
