+++ src/app/shared/views/admin/shared/guards/auth.guard.ts
@@ -8,7 +8,7 @@ import {
 
 import { Observable } from 'rxjs';
 
-import { AuthService } from './../services/auth.service';
+import { AuthService } from '../services/auth.service';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
